Decoupled base_endpoint package and move validator and converter separately, pass config as parameter instead of direct access from facade according to code review comments
