+++ common/libs/config_manager.py
@@ -1,4 +1,6 @@
 import os
+
+from common.libs.config_parser.config_dto import APIValidationDTO
 from common.libs.config_parser.config_parser import ParseConfig
 from common.libs.helpers.singleton import Singleton
 
@@ -16,3 +18,8 @@ class ConfigManager(metaclass=Singleton):
 
     def update_config(self, custom_args):
         self.config = ParseConfig(self.path_to_config, custom_args)
+
+    def get_api_validations(self) -> APIValidationDTO:
+        settings = self.config.api_settings.api_validation_settings
+        return APIValidationDTO(settings.validate_status_code, settings.validate_headers,
+                                settings.validate_body, settings.fail_if_field_is_missing)
