+++ common/libs/config_parser/config_dto.py
@@ -1,0 +1,9 @@
+from dataclasses import dataclass
+
+
+@dataclass
+class APIValidationDTO:
+    check_status_code: bool
+    check_headers: bool
+    check_body: bool
+    fail_if_field_is_missing: bool
