+++ common/rest_qa_api/tests/response_converter_test.py
@@ -3,9 +3,12 @@ import pytest
 
 from common.rest_qa_api.base_endpoint import ResponseConverterMixin
 from common.rest_qa_api.tests.tests_utils import DummyResponseBuilder, TestEndpointBuilder, exclude_fields_from_obj
+from common.rest_qa_api.tests.conftest import facade
 
 affected_fields_get = ("raw_response", "status_code", "headers", "get_data")
 affected_fields_error = ("raw_response", "status_code", "headers", "error_data")
+
+config = facade.config_manager
 
 
 def test_status_code_converter():
@@ -66,62 +69,62 @@ def test_converter_raw_response():
 
 def test_status_code_converter_as_mixin():
     orig_response = DummyResponseBuilder().method().code(500).body()
-    converted_response = TestEndpointBuilder._TestResponseModel().convert_raw_response(orig_response)
+    converted_response = TestEndpointBuilder._TestResponseModel(config).convert_raw_response(orig_response)
     assert converted_response.status_code == orig_response.status_code
     assert exclude_fields_from_obj(converted_response, affected_fields_error) == \
-           exclude_fields_from_obj(TestEndpointBuilder._TestResponseModel(), affected_fields_error)
+           exclude_fields_from_obj(TestEndpointBuilder._TestResponseModel(config), affected_fields_error)
 
 
 def test_headers_converter_as_mixin():
     orig_response = DummyResponseBuilder().method().code().body()
-    converted_response = TestEndpointBuilder._TestResponseModel().convert_raw_response(orig_response)
+    converted_response = TestEndpointBuilder._TestResponseModel(config).convert_raw_response(orig_response)
     assert converted_response.headers == orig_response.headers
     assert exclude_fields_from_obj(converted_response, affected_fields_get) == \
-           exclude_fields_from_obj(TestEndpointBuilder._TestResponseModel(), affected_fields_get)
+           exclude_fields_from_obj(TestEndpointBuilder._TestResponseModel(config), affected_fields_get)
 
 
 def test_body_json_converter_as_mixin_valid_value():
     test_body = {"testField": "testValue"}
     orig_response = DummyResponseBuilder().method().code().header({"Content-Type": "application/json"}).body(
         json.dumps(test_body))
-    converted_response = TestEndpointBuilder._TestResponseModel().convert_raw_response(orig_response)
+    converted_response = TestEndpointBuilder._TestResponseModel(config).convert_raw_response(orig_response)
     assert converted_response.get_data == test_body
     assert exclude_fields_from_obj(converted_response, affected_fields_get) == \
-           exclude_fields_from_obj(TestEndpointBuilder._TestResponseModel(), affected_fields_get)
+           exclude_fields_from_obj(TestEndpointBuilder._TestResponseModel(config), affected_fields_get)
 
 
 def test_body_json_converter_as_mixin_not_json_mime_type():
     test_body = {"testField": "testValue"}
     orig_response = DummyResponseBuilder().method().code().header({"Content-Type": "text/plain"}).body(
         json.dumps(test_body))
-    converted_response = TestEndpointBuilder._TestResponseModel().convert_raw_response(orig_response)
+    converted_response = TestEndpointBuilder._TestResponseModel(config).convert_raw_response(orig_response)
     assert converted_response.get_data == json.dumps(test_body)
     assert exclude_fields_from_obj(converted_response, affected_fields_get) == \
-           exclude_fields_from_obj(TestEndpointBuilder._TestResponseModel(), affected_fields_get)
+           exclude_fields_from_obj(TestEndpointBuilder._TestResponseModel(config), affected_fields_get)
 
 
 def test_body_json_converter_as_mixin_no_mime_type_provided():
     test_body = {"testField": "testValue"}
     orig_response = DummyResponseBuilder().method().code().body(
         json.dumps(test_body))
-    converted_response = TestEndpointBuilder._TestResponseModel().convert_raw_response(orig_response)
+    converted_response = TestEndpointBuilder._TestResponseModel(config).convert_raw_response(orig_response)
     assert converted_response.get_data == json.dumps(test_body)
     assert exclude_fields_from_obj(converted_response, affected_fields_get) == \
-           exclude_fields_from_obj(TestEndpointBuilder._TestResponseModel(), affected_fields_get)
+           exclude_fields_from_obj(TestEndpointBuilder._TestResponseModel(config), affected_fields_get)
 
 
 def test_body_json_converter_as_mixin_invalid_body():
     test_body = "testField = testValue"
     orig_response = DummyResponseBuilder().method().code().header({"Content-Type": "application/json"}).body(test_body)
     with pytest.raises(TypeError) as excinfo:
-        TestEndpointBuilder._TestResponseModel().convert_raw_response(orig_response)
+        TestEndpointBuilder._TestResponseModel(config).convert_raw_response(orig_response)
     assert f"Incorrect json format: {test_body}" in str(excinfo.value)
 
 
 def test_body_json_converter_as_mixin_error_status_code():
     test_body = "testField = testValue"
     orig_response = DummyResponseBuilder().method().code(500).body(test_body)
-    converted_response = TestEndpointBuilder._TestResponseModel().convert_raw_response(orig_response)
+    converted_response = TestEndpointBuilder._TestResponseModel(config).convert_raw_response(orig_response)
     assert converted_response.error_data == test_body
     assert exclude_fields_from_obj(converted_response, affected_fields_error) == \
-           exclude_fields_from_obj(TestEndpointBuilder._TestResponseModel(), affected_fields_error)
+           exclude_fields_from_obj(TestEndpointBuilder._TestResponseModel(config), affected_fields_error)
