+++ config/config.ini
@@ -2,7 +2,7 @@
 logdir = Logs
 log_level = INFO
 enable_libs_logging = True
-sections = api web
+sections = api
 
 [api]
 api_url = http://api.openweathermap.org/data/2.5/
